+++ app/Http/Controllers/StockController.php
@@ -1,11 +1,13 @@
 <?php
 
 namespace App\Http\Controllers;
+use RealRashid\SweetAlert\Facades\Alert;
+use Illuminate\Support\Facades\Auth;
 
 use Illuminate\Http\Request;
 use App\stock;
 use App\item;
-
+use App\User;
 use DB;
 class StockController extends Controller
 {
@@ -22,7 +24,12 @@ class StockController extends Controller
   		$stocks->item_id = $request->item_id;
   		$stocks->quantity = $request->quantity;  
 
+          if(User::where('id',Auth::id())->first()->role_id!=1) {
+            return redirect('/stock/entry')->with('message','You don\'t have permssion to add');
+          }
+          
   		$stocks->save();
+        Alert::success('Success', 'Successfully Added');
 
   		return redirect('/stock/save')->with('message','Data insert successfully.');
 
@@ -31,13 +38,18 @@ class StockController extends Controller
 	}
   public function manage(){
 
-      $stocks = DB::table('stocks')->get();
+      $stocks = DB::table('stocks')
+                  ->join('items','items.id','=','stocks.item_id')
+                  ->select('stocks.*','items.item_name')
+                  ->get();
 
       return view('admin.stock.stockManage',['stocks'=>$stocks]);
   }
 
   public function edit($id){
-
+    if(User::where('id',Auth::id())->first()->role_id!=1) {
+        return redirect('/stock/manage')->with('message','You don\'t have permssion to update');
+      }
       $stock = stock::where('id',$id)->first();
       return view('admin.stock.stockEdit',['stock'=>$stock]);
   }
@@ -58,6 +70,9 @@ class StockController extends Controller
 
   public function delete($id){
 
+    if(User::where('id',Auth::id())->first()->role_id!=1) {
+        return redirect('/stock/manage')->with('message','You don\'t have permssion to delete');
+      }
       $stockDelete = stock::find($id);
       $stockDelete->delete();
