+++ app/Http/Controllers/SuppliersDueController.php
@@ -1,11 +1,13 @@
 <?php
 
 namespace App\Http\Controllers;
+use RealRashid\SweetAlert\Facades\Alert;
 
 use Illuminate\Http\Request;
 use App\suppliers;
 use App\suppliers_dues;
 use DB;
+use App\User;
 use Illuminate\Support\Facades\Auth;
 
 class SuppliersDueController extends Controller
@@ -24,7 +26,12 @@ class SuppliersDueController extends Controller
       $suppliers->due_amount = $request->due_amount;
  
 
+      if(User::where('id',Auth::id())->first()->role_id!=1) {
+        return redirect('/suppliersDue/entry')->with('message','You don\'t have permssion to add');
+      }
+      
   		$suppliers->save();
+      Alert::success('Success', 'Successfully Added');
 
       return redirect('/suppliersDue/entry')->with('message','insert successfully');
 
@@ -33,7 +40,10 @@ class SuppliersDueController extends Controller
 
   public function manage(){
 
-      $suppliers = suppliers_dues::all();      
+      $suppliers = DB::table('suppliers_dues')   
+                  ->join('suppliers','suppliers.id','=','suppliers_dues.suppliers_id')
+                  ->select('suppliers_dues.*','suppliers.name') 
+                  ->get();       
 
 
       return view('admin.suppliersDue.suppliersManage',['suppliers'=>$suppliers]); 
@@ -54,7 +64,9 @@ class SuppliersDueController extends Controller
   public function edit($id){
 
      $suppliersId= suppliers::all();
-
+     if(User::where('id',Auth::id())->first()->role_id!=1) {
+      return redirect('/suppliersDue/manage')->with('message','You don\'t have permssion to update');
+    }
      $suppliers= suppliers_dues::where('id',$id)->first();
      return view('admin.suppliersDue.suppliersEdit',['suppliers'=>$suppliers,'suppliersId'=>$suppliersId]);
   }
@@ -83,7 +95,9 @@ class SuppliersDueController extends Controller
      if (file_exists($suppliersPic->pic)) {
        unlink($suppliersPic->pic);
      }
-     
+     if(User::where('id',Auth::id())->first()->role_id!=1) {
+      return redirect('/suppliersDue/manage')->with('message','You don\'t have permssion to delete');
+    }
      $suppliersDelete= suppliers_dues::find($id);
      $suppliersDelete->delete();
