+++ app/Http/Controllers/SuppliersPaymentController.php
@@ -1,11 +1,13 @@
 <?php
 
 namespace App\Http\Controllers;
+use RealRashid\SweetAlert\Facades\Alert;
 
 use Illuminate\Http\Request;
 use App\suppliers;
 use App\suppliers_payments;
 use DB;
+use App\User;
 use Illuminate\Support\Facades\Auth;
 
 class SuppliersPaymentController extends Controller
@@ -23,7 +25,12 @@ class SuppliersPaymentController extends Controller
       $suppliers->payment_date = $request->payment_date;
       $suppliers->payment_amount = $request->payment_amount;
  
+      if(User::where('id',Auth::id())->first()->role_id!=1) {
+        return redirect('/suppliersPayment/entry')->with('message','You don\'t have permssion to add');
+      }
+      
   		$suppliers->save();
+      Alert::success('Success', 'Successfully Added');
 
       return redirect('/suppliersPayment/entry')->with('message','insert successfully');
 
@@ -32,10 +39,23 @@ class SuppliersPaymentController extends Controller
 
   public function manage(){
 
-      $suppliers = suppliers_payments::all();      
+      $suppliers = DB::table('suppliers_payments')   
+                ->join('suppliers','suppliers.id','=','suppliers_payments.suppliers_id')
+                ->select('suppliers_payments.*','suppliers.name') 
+                ->get();       
 
       return view('admin.suppliersPayment.suppliersManage',['suppliers'=>$suppliers]); 
   }
+
+  public function search(Request $request){
+    $suppliers = DB::table('suppliers_payments')
+                ->whereBetween('payment_date', [$request->from_date,$request->to_date])
+                ->join('suppliers','suppliers.id','=','suppliers_payments.suppliers_id')
+                ->select('suppliers_payments.*','suppliers.name')
+                ->get();
+
+        return view('admin.suppliersPayment.suppliersManage',['suppliers'=>$suppliers]); 
+    }
 
 
   public function singlesuppliers($id){
@@ -50,6 +70,9 @@ class SuppliersPaymentController extends Controller
   }
 
   public function edit($id){
+    if(User::where('id',Auth::id())->first()->role_id!=1) {
+      return redirect('/suppliersPayment/manage')->with('message','You don\'t have permssion to update');
+    }
 
      $suppliers= suppliers_payments::where('id',$id)->first();
      return view('admin.suppliersPayment.suppliersEdit',['suppliers'=>$suppliers]);
@@ -79,8 +102,9 @@ class SuppliersPaymentController extends Controller
      if (file_exists($suppliersPic->pic)) {
        unlink($suppliersPic->pic);
      }
-     
-
+     if(User::where('id',Auth::id())->first()->role_id!=1) {
+      return redirect('/suppliersPayment/manage')->with('message','You don\'t have permssion to update');
+    }
 
      $suppliersDelete= suppliers_payments::find($id);
      $suppliersDelete->delete();
